requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=2.0.0
langchain-community>=0.0.10
google-genai>=1.0.0
yfinance>=0.2.32
nltk>=3.8.1
wikipedia-api>=0.6.0
pandas>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0

